working with variadic function

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -15,7 +15,7 @@ int sum_them_all(const unsigned int n, ...)
 	if (n == 0)
 		return (0);
 	va_start(args, n);
-	for (i = 0; i < n; i++)
+	for (i = 0; i < (int)n; i++)
 	{
 		total += total + va_arg(args, int);
 	}
